usersコントローラーの追加とユーザー関連を修正
