+++ app/controllers/master_controller.rb
@@ -1,45 +1,19 @@
 # frozen_string_literal: true
 
 class MasterController < ApplicationController
-  def index
-    @users = User.page(params[:page]).per(10)
-  end
+  def index; end
 
-  def new
-    @user = User.new
-  end
+  def new; end
 
-  def create
-    @user = User.new(user_params)
-    if @user.save
-      redirect_to master_index_path
-      flash[:success] = 'ユーザを登録しました'
-    else
-      render :new
-    end
-  end
+  def create; end
 
-  def show
-    @user = User.find(params[:id])
-  end
+  def show; end
 
-  def edit
-    @user = User.find(params[:id])
-  end
+  def edit; end
 
-  def update
-    @user = User.find(params[:id])
-    if @user.update(user_params)
-      redirect_to master_index_path
-      flash[:success] = 'ユーザを更新しました'
-    else
-      render :edit
-    end
-  end
+  def update; end
 
   private
 
-  def user_params
-    params.require(:user).permit(:name, :email, :password, :role, :icon, :organization_id)
-  end
+
 end
